IDEMPIERE-6570 Improve attachment support for large file
- Implement DB Large Object support as a new storage method "DB_LOB"
- Fix SaveAllAsZip not working for file > 2GB
- Fix migration of storage provider not working for file > 2GB
- Fix FileSystem storage provider not working for file > 2GB
- Use zip compression for DB_LOB
