+++ org.adempiere.base/OSGI-INF/attachmentdb_lob.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.compiere.model.AttachmentDBLOB">
+   <implementation class="org.compiere.model.AttachmentDBLOB"/>
+   <service>
+      <provide interface="org.compiere.model.IAttachmentStore"/>
+   </service>
+   <property name="method" type="String" value="DB_LOB"/>
+</scr:component>
