+++ org.adempiere.base/src/org/compiere/model/I_AD_StorageProvider.java
@@ -22,7 +22,7 @@ import org.compiere.util.KeyNamePair;
 
 /** Generated Interface for AD_StorageProvider
  *  @author iDempiere (generated) 
- *  @version Release 12
+ *  @version Release 13
  */
 public interface I_AD_StorageProvider 
 {
